refact : controller 패키지 경로 변경 및 기존 api spec test 제거
