+++ src/main/java/personal/yejin/foodDelivery/delivery/dto/DeliveryStatusUpdateResponse.java
@@ -1,6 +1,8 @@
 package personal.yejin.foodDelivery.delivery.dto;
 
+import personal.yejin.foodDelivery.delivery.model.DeliveryStatus;
+
 public record DeliveryStatusUpdateResponse(
         long orderId,
-        DeliveryStatusUpdateRequest.DeliveryStatus updatedStatus
+        DeliveryStatus updatedStatus
 ){}
