+++ src/main/java/personal/yejin/foodDelivery/delivery/dto/DispatchRequest.java
@@ -1,7 +1,6 @@
 package personal.yejin.foodDelivery.delivery.dto;
 
+import personal.yejin.foodDelivery.delivery.model.DeliveryType;
+
 public record DispatchRequest(DeliveryType deliveryType) {
-    public enum DeliveryType {
-        SINGLE, BUNDLE
-    }
 }
