+++ src/main/java/personal/yejin/foodDelivery/delivery/dto/DispatchResponse.java
@@ -1,14 +1,13 @@
 package personal.yejin.foodDelivery.delivery.dto;
 
+import personal.yejin.foodDelivery.delivery.model.DeliveryType;
+import personal.yejin.foodDelivery.delivery.model.DispatchStatus;
+
 public record DispatchResponse(
         long orderId,
         DispatchStatus dispatchStatus,
-        DispatchRequest.DeliveryType deliveryType
+        DeliveryType deliveryType
 ) {
 
-    public enum DispatchStatus {
-        REQUESTED,
-        ACCEPTED,
-        REJECTED
-    }
+
 }
