+++ src/main/java/personal/yejin/foodDelivery/delivery/model/DeliveryStatus.java
@@ -1,0 +1,7 @@
+package personal.yejin.foodDelivery.delivery.model;
+
+public enum DeliveryStatus {
+    PICKED_UP,
+    DELIVERED,
+    CANCEL
+}
