+++ src/main/java/personal/yejin/foodDelivery/delivery/model/DeliveryType.java
@@ -1,0 +1,5 @@
+package personal.yejin.foodDelivery.delivery.model;
+
+public enum DeliveryType {
+    SINGLE, BUNDLE
+}
