+++ src/main/java/personal/yejin/foodDelivery/delivery/model/DispatchStatus.java
@@ -1,0 +1,7 @@
+package personal.yejin.foodDelivery.delivery.model;
+
+public enum DispatchStatus {
+    REQUESTED,
+    ACCEPTED,
+    REJECTED
+}
