+++ src/main/java/personal/yejin/foodDelivery/payment/dto/DiscountDetails.java
@@ -1,0 +1,6 @@
+package personal.yejin.foodDelivery.payment.dto;
+
+public record DiscountDetails(
+        int couponDiscount,
+        int baeminClubDiscount
+) {}
