--- src/main/java/personal/yejin/foodDelivery/rider/dto/CurrentRouteResponse.java
@@ -1,19 +1,0 @@
-package personal.yejin.foodDelivery.rider.dto;
-
-
-import java.util.List;
-public record CurrentRouteResponse (
-    List<Stop> stops
-)
-{
-    public record Stop (
-        int sequence,
-        StopType type,
-        long orderId,
-        String address
-    ){}
-
-    public enum StopType {
-        PICKUP, DELIVERY
-    }
-}
