+++ src/main/java/personal/yejin/foodDelivery/rider/dto/RiderLocationRequest.java
@@ -2,4 +2,5 @@ package personal.yejin.foodDelivery.rider.dto;
 
 public record RiderLocationRequest(
         double latitude,
-        double longitude){}
+        double longitude
+){}
