+++ src/main/java/personal/yejin/foodDelivery/route/dto/CurrentRouteResponse.java
@@ -1,0 +1,8 @@
+package personal.yejin.foodDelivery.route.dto;
+
+
+import java.util.List;
+public record CurrentRouteResponse (
+    List<Stop> stops
+)
+{ }
