+++ src/main/java/personal/yejin/foodDelivery/route/dto/Stop.java
@@ -1,0 +1,10 @@
+package personal.yejin.foodDelivery.route.dto;
+
+import personal.yejin.foodDelivery.route.model.StopType;
+
+public record Stop (
+        int sequence,
+        StopType type,
+        long orderId,
+        String address
+){}
