+++ src/main/java/personal/yejin/foodDelivery/route/model/StopType.java
@@ -1,0 +1,5 @@
+package personal.yejin.foodDelivery.route.model;
+
+public enum StopType {
+    PICKUP, DELIVERY
+}
